moved data processing functions in pp3 to dataprocessing
